requirements.txt
crewai
python-dotenv
langchain-openai